requirements.txt
tabulate==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
pyzmq==26.0.3
requests==2.32.3
urllib3==2.2.2
